+++ src/engine/search/tables.rs
@@ -1,7 +1,36 @@
 use crate::chess::moves::Move;
 use crate::chess::player::Player;
 use crate::chess::square::Square;
-use crate::engine::search::move_ordering;
+use crate::engine::search::{move_ordering, MAX_SEARCH_DEPTH_SIZE};
+
+pub struct KillersTable([[Option<Move>; 2]; MAX_SEARCH_DEPTH_SIZE]);
+
+impl KillersTable {
+    pub const fn new() -> Self {
+        Self([[None; 2]; MAX_SEARCH_DEPTH_SIZE])
+    }
+
+    pub fn get_0(&self, plies: u8) -> Option<Move> {
+        let plies = plies as usize;
+        self.0[plies][0]
+    }
+
+    pub fn get_1(&self, plies: u8) -> Option<Move> {
+        let plies = plies as usize;
+        self.0[plies][1]
+    }
+
+    pub fn try_push(&mut self, plies: u8, mv: Move) {
+        let plies = plies as usize;
+
+        let killer_1 = self.0[plies][0];
+
+        if Some(mv) != killer_1 {
+            self.0[plies][1] = killer_1;
+            self.0[plies][0] = Some(mv);
+        }
+    }
+}
 
 pub struct HistoryTable([[[i32; Square::N]; Square::N]; Player::N]);